requirement.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
idna==3.4
imageio-ffmpeg==0.4.3
importlib-metadata==5.0.0
ninja==1.10.2.4
numpy==1.21.6
opencv-python==4.6.0.66
Pillow==9.2.0
pyspng==0.1.0
requests==2.28.1
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.9.0
